requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0

# Google AI
google-generativeai>=0.8.0

# Vector Database & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.0

# Web Framework
streamlit>=1.30.0
plotly>=5.18.0
streamlit-aggrid>=0.3.4

# ML & Analytics
scikit-learn>=1.4.0
scipy>=1.11.0

# Data Processing
pydantic>=2.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Storage
sqlalchemy>=2.0.0

# Utilities
tenacity>=8.2.0
aiohttp>=3.9.0
openpyxl>=3.1.0

# Testing
pytest>=7.4.0
